fixed model datatypes once and for all

+++ lib/ui/driver_earning/driver_earning_page.dart
@@ -51,7 +51,7 @@ class _DriverEarningPageState extends State<DriverEarningPage> {
   Widget _buildContent(BuildContext context, EarningModel? earning) {
     final colorScheme = Theme.of(context).colorScheme;
     if (earning != null) {
-      amount = earning.totalEarning;
+      amount = earning.totalEarning.toInt();
     }
     return Container(
       padding: EdgeInsets.fromLTRB(0, 20.0, 0, 20.0),
